# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DevOps Forge API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ModuleCategory = typeof ModuleCategory[keyof typeof ModuleCategory];


export const ModuleCategory = {
  containers: 'containers',
  ci_cd: 'ci_cd',
  cloud: 'cloud',
  infrastructure: 'infrastructure',
  orchestration: 'orchestration',
} as const;

export interface Module {
  id: number;
  slug: string;
  title: string;
  description: string;
  icon: string;
  category: ModuleCategory;
  totalChallenges: number;
  completedChallenges: number;
  isUnlocked: boolean;
  estimatedHours?: number;
}

export type ChallengeDifficulty = typeof ChallengeDifficulty[keyof typeof ChallengeDifficulty];


export const ChallengeDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface Challenge {
  id: number;
  moduleId: number;
  title: string;
  description?: string;
  difficulty: ChallengeDifficulty;
  isCompleted: boolean;
  xpReward: number;
  tags?: string[];
}

export interface ModuleDetail {
  id: number;
  slug: string;
  title: string;
  description: string;
  icon: string;
  category: string;
  isUnlocked: boolean;
  challenges: Challenge[];
}

export type ChallengeDetailDifficulty = typeof ChallengeDetailDifficulty[keyof typeof ChallengeDetailDifficulty];


export const ChallengeDetailDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface ChallengeDetail {
  id: number;
  moduleId: number;
  title: string;
  description?: string;
  difficulty: ChallengeDetailDifficulty;
  isCompleted: boolean;
  xpReward: number;
  content: string;
  hints: string[];
  tags?: string[];
}

export interface Progress {
  challengeId: number;
  isCompleted: boolean;
  completedAt: string;
}

export interface ProgressSummary {
  totalXp: number;
  level: number;
  levelName?: string;
  completedChallenges: number;
  totalChallenges: number;
  streakDays: number;
  modulesStarted: number;
  modulesCompleted: number;
  nextLevelXp?: number;
  currentLevelXp?: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  challenge_completed: 'challenge_completed',
  module_started: 'module_started',
  badge_earned: 'badge_earned',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  xpEarned: number;
  /** @nullable */
  moduleName?: string | null;
  occurredAt: string;
}

export interface DocTopic {
  id: number;
  moduleId: number;
  title: string;
  summary: string;
  readingMinutes: number;
  tags?: string[];
}

export interface DocTopicDetail {
  id: number;
  moduleId: number;
  title: string;
  summary: string;
  content: string;
  readingMinutes: number;
  tags?: string[];
}

export type ListChallengesParams = {
moduleId?: number;
difficulty?: ListChallengesDifficulty;
};

export type ListChallengesDifficulty = typeof ListChallengesDifficulty[keyof typeof ListChallengesDifficulty];


export const ListChallengesDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export type ListDocTopicsParams = {
moduleId?: number;
};