# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DevOps Forge API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: modules
    description: Learning module operations
  - name: challenges
    description: Challenge operations
  - name: progress
    description: User progress operations
  - name: docs
    description: Documentation operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /modules:
    get:
      operationId: listModules
      tags: [modules]
      summary: List all learning modules
      responses:
        "200":
          description: List of modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"

  /modules/{id}:
    get:
      operationId: getModule
      tags: [modules]
      summary: Get a module with its challenges
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Module detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleDetail"
        "404":
          description: Not found

  /challenges:
    get:
      operationId: listChallenges
      tags: [challenges]
      summary: List challenges, optionally filtered by module
      parameters:
        - name: moduleId
          in: query
          required: false
          schema:
            type: integer
        - name: difficulty
          in: query
          required: false
          schema:
            type: string
            enum: [beginner, intermediate, advanced]
      responses:
        "200":
          description: List of challenges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Challenge"

  /challenges/{id}:
    get:
      operationId: getChallenge
      tags: [challenges]
      summary: Get a single challenge
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Challenge detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeDetail"
        "404":
          description: Not found

  /challenges/{id}/complete:
    post:
      operationId: completeChallenge
      tags: [challenges]
      summary: Mark a challenge as completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Progress"

  /progress/summary:
    get:
      operationId: getProgressSummary
      tags: [progress]
      summary: Get overall learning progress summary for dashboard
      responses:
        "200":
          description: Progress summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressSummary"

  /progress/activity:
    get:
      operationId: getActivityFeed
      tags: [progress]
      summary: Get recent learning activity
      responses:
        "200":
          description: Recent activity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  /docs/topics:
    get:
      operationId: listDocTopics
      tags: [docs]
      summary: List documentation topics
      parameters:
        - name: moduleId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of topics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocTopic"

  /docs/topics/{id}:
    get:
      operationId: getDocTopic
      tags: [docs]
      summary: Get a documentation topic
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Topic detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocTopicDetail"
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Module:
      type: object
      required: [id, slug, title, description, icon, category, totalChallenges, completedChallenges, isUnlocked]
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        category:
          type: string
          enum: [containers, ci_cd, cloud, infrastructure, orchestration]
        totalChallenges:
          type: integer
        completedChallenges:
          type: integer
        isUnlocked:
          type: boolean
        estimatedHours:
          type: integer

    ModuleDetail:
      type: object
      required: [id, slug, title, description, icon, category, challenges, isUnlocked]
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        category:
          type: string
        isUnlocked:
          type: boolean
        challenges:
          type: array
          items:
            $ref: "#/components/schemas/Challenge"

    Challenge:
      type: object
      required: [id, moduleId, title, difficulty, isCompleted, xpReward]
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        isCompleted:
          type: boolean
        xpReward:
          type: integer
        tags:
          type: array
          items:
            type: string

    ChallengeDetail:
      type: object
      required: [id, moduleId, title, difficulty, isCompleted, xpReward, content, hints]
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        isCompleted:
          type: boolean
        xpReward:
          type: integer
        content:
          type: string
        hints:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string

    Progress:
      type: object
      required: [challengeId, isCompleted, completedAt]
      properties:
        challengeId:
          type: integer
        isCompleted:
          type: boolean
        completedAt:
          type: string

    ProgressSummary:
      type: object
      required: [totalXp, level, completedChallenges, totalChallenges, streakDays, modulesStarted, modulesCompleted]
      properties:
        totalXp:
          type: integer
        level:
          type: integer
        levelName:
          type: string
        completedChallenges:
          type: integer
        totalChallenges:
          type: integer
        streakDays:
          type: integer
        modulesStarted:
          type: integer
        modulesCompleted:
          type: integer
        nextLevelXp:
          type: integer
        currentLevelXp:
          type: integer

    ActivityItem:
      type: object
      required: [id, type, title, xpEarned, occurredAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [challenge_completed, module_started, badge_earned]
        title:
          type: string
        xpEarned:
          type: integer
        moduleName:
          type: ["string", "null"]
        occurredAt:
          type: string

    DocTopic:
      type: object
      required: [id, moduleId, title, summary, readingMinutes]
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        summary:
          type: string
        readingMinutes:
          type: integer
        tags:
          type: array
          items:
            type: string

    DocTopicDetail:
      type: object
      required: [id, moduleId, title, summary, content, readingMinutes]
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        summary:
          type: string
        content:
          type: string
        readingMinutes:
          type: integer
        tags:
          type: array
          items:
            type: string