# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DevOps Forge API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemType } from './activityItemType';

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  xpEarned: number;
  /** @nullable */
  moduleName?: string | null;
  occurredAt: string;
}