# lib/api-zod/src/generated/types/challenge.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DevOps Forge API
 * OpenAPI spec version: 0.1.0
 */
import type { ChallengeDifficulty } from './challengeDifficulty';

export interface Challenge {
  id: number;
  moduleId: number;
  title: string;
  description?: string;
  difficulty: ChallengeDifficulty;
  isCompleted: boolean;
  xpReward: number;
  tags?: string[];
}