# lib/api-zod/src/generated/types/challengeDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DevOps Forge API
 * OpenAPI spec version: 0.1.0
 */
import type { ChallengeDetailDifficulty } from './challengeDetailDifficulty';

export interface ChallengeDetail {
  id: number;
  moduleId: number;
  title: string;
  description?: string;
  difficulty: ChallengeDetailDifficulty;
  isCompleted: boolean;
  xpReward: number;
  content: string;
  hints: string[];
  tags?: string[];
}