# lib/api-zod/src/generated/types/docTopicDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DevOps Forge API
 * OpenAPI spec version: 0.1.0
 */

export interface DocTopicDetail {
  id: number;
  moduleId: number;
  title: string;
  summary: string;
  content: string;
  readingMinutes: number;
  tags?: string[];
}