# lib/api-zod/src/generated/types/module.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DevOps Forge API
 * OpenAPI spec version: 0.1.0
 */
import type { ModuleCategory } from './moduleCategory';

export interface Module {
  id: number;
  slug: string;
  title: string;
  description: string;
  icon: string;
  category: ModuleCategory;
  totalChallenges: number;
  completedChallenges: number;
  isUnlocked: boolean;
  estimatedHours?: number;
}