# lib/api-zod/src/generated/types/moduleDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DevOps Forge API
 * OpenAPI spec version: 0.1.0
 */
import type { Challenge } from './challenge';

export interface ModuleDetail {
  id: number;
  slug: string;
  title: string;
  description: string;
  icon: string;
  category: string;
  isUnlocked: boolean;
  challenges: Challenge[];
}