# lib/api-zod/src/generated/types/progress.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DevOps Forge API
 * OpenAPI spec version: 0.1.0
 */

export interface Progress {
  challengeId: number;
  isCompleted: boolean;
  completedAt: string;
}