# lib/api-zod/src/generated/types/progressSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DevOps Forge API
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressSummary {
  totalXp: number;
  level: number;
  levelName?: string;
  completedChallenges: number;
  totalChallenges: number;
  streakDays: number;
  modulesStarted: number;
  modulesCompleted: number;
  nextLevelXp?: number;
  currentLevelXp?: number;
}